# swagger.yaml
openapi: 3.0.0
info:
  title: 4TP API
  description: 4TP API
  version: 0.0.1
servers:
  - url: "https://for-the-players.herokuapp.com/"
...
paths:
...
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: The ID of the user
        schema:
          $ref: "#/components/schemas/MongoID"
    get:
      tags:
        - user
      security:
        - BearerToken: [ ]
      summary: Returns a single user's details
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "The User"
          content:
            application/json:
              schema:
                type: object
                properties:
                  elo:
                    type: integer
                    description: The player's ELO
                  email:
                    type: string
                    format: email
                    description: "The user's email"
                  username:
                    type: string
                    description: "The user's username"
                  createdAt:
                    type: string
                    format: date
                  platforms:
                    type: array
                    description: "The platforms to which the user has registered"
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: "The platform's name"
                        username:
                          type: string
                          description: "The user's username for this platform"
                  tournaments:
                    type: array
                    description: "The tournament of which the user is a part of"
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: The tournament's name
                        game:
                          type: string
                          description: The name of the tournament's game
                        finished:
                          type: boolean
                          description: If the tournament is over
                        platform:
                          type: string
                          description: The name of the tournament's platform
                        type:
                          $ref: "#/components/schemas/TournamentTypes"
                        team:
                          type: object
                          description: The user's team in that tournament
                          properties:
                            name:
                              type: string
                              description: The team's name
                  invites:
                    type: array
                    description: List of invites the user has received
                    items:
                      type: object
                      properties:
                        tournamentId:
                          type: string
                          description: The ID of the tournament this invite refres to
                        teamName:
                          type: string
                          description: The name of the team the user got invited to
                        status:
                          $ref: "#/components/schemas/TeamInvites"
                  coins:
                    type: integer
                    description: The amount of tokens a user has
    patch:
      summary: Updates the user profile
      security:
        - BearerAuth: [ ]
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              required:
                - oldPassword
                - newPassword
              type: object
              properties:
                oldPassword:
                  description: Old password
                  type: string
                newPassword:
                  description: New password
                  type: string
                username:
                  description: New username
                  type: string
                platforms:
                  description: List of platforms and the logged user's usernames
                  type: array
                  items:
                    type: object
                    properties:
                      username:
                        type: string
                        description: The user's username
                      _id:
                        $ref: "#/components/schemas/MongoID"
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "User updated"
  /tournaments:
    get:
      summary: List all visible tournaments
      security:
        - BearerAuth: [ ]
      tags:
        - tournaments
      operationId: listVisibleTournaments
      parameters:
        - in: query
          name: gameId
          schema:
            $ref: "#/components/schemas/MongoID"
          description: Only show tournaments for a given game
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/TournamentTypes"
          description: Only show a type of tournament
      responses:
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "List of all visible tournaments"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: "The name of the tournament"
                    _id:
                      type: string
                      description: The Mongo ID of the tournament
                    startsOn:
                      type: string
                      format: date-time
                      description: The time on which the tournament starts
                    endsOn:
                      type: string
                      format: date-time
                      description: The time on which the tournament ends
                    rulesets:
                      type: array
                      items:
                        type: string
                        description: The mongo ID of the ruleset
                    type:
                      $ref: "#/components/schemas/TournamentTypes"
                    numberOfTeams:
                      type: integer
                      description: How many teams are already signed up to this tournament
                    imgur:
                      type: string
                      description: The URL of the tournament's image
                    game:
                      type: string
                      description: The name of the game this tournament is on
                    open:
                      type: boolean
                      description: If new teams can join the tournament
                    challongeId:
                      type: integer
                      description: The Challonge ID of the tournament
  /tournaments/{tournamentId}:
    parameters:
      - name: tournamentId
        in: path
        required: true
        description: The MongoDB ID of the tournament
        schema:
          $ref: "#/components/schemas/MongoID"
    get:
      summary: Get a tournament details
      security:
        - BearerAuth: [ ]
      tags:
        - tournaments
      operationId: getTeam
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "Something was not found, check errorMessage for further info"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: "The name of the tournament"
                  _id:
                    type: string
                    description: The Mongo ID of the tournament
                  startsOn:
                    type: string
                    format: date-time
                    description: The time on which the tournament starts
                  endsOn:
                    type: string
                    format: date-time
                    description: The time on which the tournament ends
                  rulesets:
                    type: array
                    description: The list of rulesets used in this tournament
                    items:
                      type: object
                      properties:
                        _id:
                          $ref: "#/components/schemas/MongoID"
                        maps:
                          type: array
                          items:
                            type: string
                        maxNumberOfPlayersPerTeam:
                          type: integer
                        minNumberOfPlayersPerTeam:
                          type: integer
                        description:
                          type: string
                        name:
                          type: string
                        bestOf:
                          type: integer
                  type:
                    $ref: "#/components/schemas/TournamentTypes"
                  teams:
                    type: array
                    description: The teams signed up to this tournament
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: The ID of the team
                        name:
                          type: string
                          description: The team's name
                        elo:
                          type: integer
                          description: The team's ELO
                        points:
                          type: integer
                          description: The team's ELO
                        members:
                          type: array
                          description: The members of the team
                          items:
                            type: object
                            properties:
                              role:
                                $ref: "#/components/schemas/TeamRoles"
                              userId:
                                type: string
                                minLength: 24
                                maxLength: 24
                                description: The Mongo ID of the user that's a part of this team
                        teamStatus:
                          $ref: "#/components/schemas/TeamStatuses"
                        wins:
                          type: integer
                          description: The total wins for this team in this tournament
                        losses:
                          type: integer
                          description: The total losses for this team in this tournament
                        ties:
                          type: integer
                          description: The total ties for this team in this tournament
                  userTeam:
                    description: The team of the user, if present
                    type: object
                    properties:
                      name:
                        type: string
                        description: The team's name
                      elo:
                        type: integer
                        description: The team's ELO
                      members:
                        type: array
                        description: The members of the team
                        items:
                          type: object
                          properties:
                            role:
                              $ref: "#/components/schemas/TeamRoles"
                            userId:
                              type: string
                              minLength: 24
                              maxLength: 24
                              description: The Mongo ID of the user that's a part of this team
                  imgur:
                    type: string
                    description: The URL of the tournament's image
                  game:
                    type: string
                    description: The name of the game this tournament is on
                  matches:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: The ID of the match
                      teamOne:
                        type: object
                        description: The team that created the match
                        properties:
                          _id:
                            $ref: "#/components/schemas/MongoID"
                          name:
                            type: string
                          result:
                            $ref: "#/components/schemas/TeamSubmittedMatchResults"
                      teamTwo:
                        type: object
                        description: If present is the team that accepted the match
                        properties:
                          _id:
                            $ref: "#/components/schemas/MongoID"
                          name:
                            type: string
                          result:
                            $ref: "#/components/schemas/TeamSubmittedMatchResults"
                      status:
                        $ref: "#/components/schemas/MatchStatuses"
                      createdAt:
                        type: string
                        format: date-time
                        description: "When the match was first created"
                      acceptedAt:
                        type: string
                        format: date-time
                        description: "When the match was accepted"
                  open:
                    type: boolean
                    description: If new teams can join the tournament
                  minTeamSizePerMatch:
                    type: integer
                  maxTeamSizePerMatch:
                    type: integer
                  rules:
                    type: string
                    description: Tournament specific rules. Different from the ruleset
                  challongeId:
                    type: integer
                    description: The Challonge ID of the tournament
...
  /admin/tournaments/{tournamentId}:
    parameters:
      - name: tournamentId
        in: path
        required: true
        description: The MongoDB ID of the tournament
        schema:
          $ref: "#/components/schemas/MongoID"
    patch:
      security:
        - BearerAuth: [ ]
      tags:
        - admin
      operationId: updateTournament
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                game:
                  type: string
                platform:
                  type: string
                show:
                  type: boolean
                open:
                  type: boolean
                startsOn:
                  type: string
                  format: date-time
                endsOn:
                  type: string
                  format: date-time
                rulesets:
                  type: array
                  items:
                    type: string
                    description: ruleset ID
                    minLength: 24
                    maxLength: 24
                type:
                  $ref: "#/components/schemas/TournamentTypes"
                imgUrl:
                  description: >
                    The imgurl link of the new team's image.
                    If this is present `imgBase64` will be ignored
                  type: string
                imgBase64:
                  description: >
                    The base64 of the new team's image.
                    If `imgUrl` is present this will be ignored
                  type: string
                minTeamSizePerMatch:
                  type: integer
                maxTeamSizePerMatch:
                  type: integer
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "Resource not found"
        "200":
          description: "Tournament updated"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    SchemaValidationError:
      type: array
      items:
        type: object
        properties:
          value:
            type: string
            description: "The value that was received by the server"
          msg:
            type: string
            description: "Error message"
          param:
            type: string
            description: "Name of the field that caused the error"
          location:
            type: string
            description: "Where the field was that was received by the server"
    ErrorMessageResponse:
      type: object
      properties:
        errorMessage:
          type: string
...
    MatchStatuses:
      description: "The possible statuses of a match"
      type: string
      enum:
        - PENDING
        - TIE
        - TEAM1
        - TEAM2
        - DISPUTE
    TeamSubmittedMatchResults:
      description: "The list of possible results a team can choose from"
      type: string
      enum:
        - WIN
        - LOSS
        - TIE
...
    TeamRoles:
      description: "The possible roles for team members"
      type: string
      enum:
        - MEMBER
        - LEADER
    TeamInvites:
      description: "The possible invite statuses for team members"
      type: string
      enum:
        - PENDING
        - DECLINED
        - ACCEPTED
    TournamentTypes:
      description: "The possible types of a tournament"
      type: string
      enum:
        - TOURNAMENT
        - LADDER
        - BRACKET
    TeamStatuses:
      description: " Values that are added to a team object at runtime to tell the frontend if the team at it's current state can play in the tournament or not"
      type: string
      enum:
        - OK
        - NOT OK
    MongoID:
      type: string
      maxLength: 24
      minLength: 24